Refactor: table -> order service Layer 연관관계 제거
