+++ src/main/java/kitchenpos/order/application/OrderTableService.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.application;
+
+public interface OrderTableService {
+    boolean isTableEmpty(Long tableId);
+}
